reboot-router/requirements.txt
selenium==4.35.0
python-dotenv==1.2.2
